requirements.txt
python-dotenv==1.0.0
selenium==4.15.2
undetected-chromedriver>=3.5.5
setuptools>=80.9.0